feat: milkCTRL render refactor, control-action layer, and diagnostics
