+++ src/cli/overview/overview_render.c
@@ -115,307 +115,6 @@ static void ov_apply_rank_sort(OV_MODEL *mm)
 
 void render_pad_spaces(int chars_written, int panel_width);
 
-/**
- * render_highlighted_name - print a string with regex match highlighting
- */
-void render_highlighted_name(
-    const char *name,
-    int         max_len,
-    regex_t    *re,
-    int         has_re,
-    ov_rgb_t    normal_fg,
-    ov_rgb_t    row_bg)
-{
-    int len = (int) strlen(name);
-    if (len > max_len) len = max_len;
-
-    regmatch_t pm[1];
-    if (has_re && regexec(re, name, 1, pm, 0) == 0)
-    {
-        int b_len = pm[0].rm_so;
-        if (b_len > max_len) b_len = max_len;
-
-        int m_len = pm[0].rm_eo - pm[0].rm_so;
-        if (b_len + m_len > max_len) m_len = max_len - b_len;
-
-        int a_len = len - (b_len + m_len);
-
-        if (b_len > 0)
-        {
-            ov_buf_printf("%.*s", b_len, name);
-        }
-        if (m_len > 0)
-        {
-            ov_buf_bold();
-            ov_buf_fg(255, 255, 255);
-            ov_buf_printf("%.*s", m_len, name + b_len);
-            ov_buf_reset_attr();
-            ov_theme_bg(row_bg);
-            ov_theme_fg(normal_fg);
-        }
-        if (a_len > 0)
-        {
-            ov_buf_printf("%.*s", a_len, name + b_len + m_len);
-        }
-    }
-    else
-    {
-        ov_buf_printf("%.*s", len, name);
-    }
-
-    int pad = max_len - len;
-    if (pad > 0)
-    {
-        ov_buf_hline(' ', pad);
-    }
-}
-
-/**
- * ov_filter_build - build filtered index array.
- * @pattern:  regex pattern string (empty = match all)
- * @names:    array of name pointers
- * @count:    total item count
- * @out:      output index array (caller-allocated)
- * @max_out:  capacity of @out
- *
- * Return: number of matching indices written to @out.
- */
-int ov_filter_build(
-    const char  *pattern,
-    const char **names,
-    int          count,
-    int         *out,
-    int          max_out)
-{
-    if (pattern[0] == '\0')
-    {
-        /* No filter — all items match */
-        int n = count < max_out ? count : max_out;
-        for (int i = 0; i < n; i++)
-        {
-            out[i] = i;
-        }
-        return n;
-    }
-
-    regex_t re;
-    if (regcomp(&re, pattern,
-                REG_EXTENDED | REG_NOSUB
-                | REG_ICASE) != 0)
-    {
-        /* Invalid regex — show all */
-        int n = count < max_out ? count : max_out;
-        for (int i = 0; i < n; i++)
-        {
-            out[i] = i;
-        }
-        return n;
-    }
-
-    int n = 0;
-    for (int i = 0; i < count && n < max_out; i++)
-    {
-        if (regexec(&re, names[i], 0, NULL, 0) == 0)
-        {
-            out[n++] = i;
-        }
-    }
-    regfree(&re);
-    return n;
-}
-
-/* =========================================================
- * Cross-panel relation highlight
- * ========================================================= */
-
-/*
- * Bitset helpers and OV_RELATED defined in
- * overview_render_internal.h
- */
-
-void bset(uint64_t *words, int idx)
-{
-    words[idx / BITS_PER_WORD] |= (UINT64_C(1) << (idx % BITS_PER_WORD));
-}
-
-int bget(const uint64_t *words, int idx)
-{
-    return (words[idx / BITS_PER_WORD] >> (idx % BITS_PER_WORD)) & 1;
-}
-
-/* =========================================================
- * Stream lineage (ancestors / descendants)
- * ========================================================= */
-
-/* Lineage types and BFS now in stream_graph.{h,c} */
-
-/**
- * ov_compute_related - find items in other panels related to the selection.
- * @lay: current layout (holds focus + sel_* indices)
- * @m:   data model
- * @out: result bitsets (cleared on entry)
- *
- * Traverses graph edges from the selected node and marks all
- * reachable streams, FPS entries, and processes.
- */
-static void ov_compute_related(
-    const OV_LAYOUT *lay,
-    const OV_MODEL  *m,
-    OV_RELATED      *out)
-{
-    memset(out, 0, sizeof(*out));
-    /* fps_param_mask initialised to 0 by memset — no matches yet */
-
-    /* Use frozen selection when freeze is active */
-    ov_focus_t focus   = lay->freeze
-                         ? lay->freeze_focus
-                         : lay->focus;
-    int sel_stream_idx = lay->freeze
-                         ? lay->freeze_sel_stream
-                         : lay->sel_stream;
-    int sel_proc_idx   = lay->freeze
-                         ? lay->freeze_sel_proc
-                         : lay->sel_proc;
-    int sel_fps_idx    = lay->freeze
-                         ? lay->freeze_sel_fps
-                         : lay->sel_fps;
-
-    /* Determine the graph node index of the selected item */
-    int sel_node = -1;
-    if (focus == OV_FOCUS_STREAMS && sel_stream_idx >= 0
-        && sel_stream_idx < m->nb_streams)
-    {
-        sel_node = m->streams[sel_stream_idx].node_idx;
-    }
-    else if (focus == OV_FOCUS_FPS && sel_fps_idx >= 0
-             && sel_fps_idx < m->nb_fps)
-    {
-        sel_node = m->fps[sel_fps_idx].node_idx;
-    }
-    else if (focus == OV_FOCUS_PROCS && sel_proc_idx >= 0
-             && sel_proc_idx < m->nb_procs)
-    {
-        sel_node = m->procs[sel_proc_idx].node_idx;
-        out->sel_pid = m->procs[sel_proc_idx].PID;
-    }
-
-    if (sel_node < 0)
-    {
-        return;
-    }
-
-    /* Walk all edges; mark neighbours of sel_node */
-    for (int ei = 0; ei < m->nb_edges; ei++)
-    {
-        const OV_EDGE *e = &m->edges[ei];
-        int other    = -1;
-        int is_write = 0; /* 1 = proc writes stream */
-        /* For FPS edges: which end is the stream node? */
-        int fps_is_src = 0;
-
-        if (e->src_node == sel_node)
-        {
-            other      = e->tgt_node;
-            is_write   = (e->type == OV_EDGE_PROC_WRITES_STREAM);
-            fps_is_src = 0; /* FPS is tgt when stream→FPS */
-        }
-        else if (e->tgt_node == sel_node)
-        {
-            other      = e->src_node;
-            is_write   = 0;
-            fps_is_src = 1; /* FPS is src when FPS→stream */
-        }
-
-        if (other < 0 || other >= m->nb_nodes)
-        {
-            continue;
-        }
-
-        const OV_NODE *n = &m->nodes[other];
-        if (n->type == OV_NODE_STREAM && n->index >= 0
-            && n->index < m->nb_streams)
-        {
-            bset(out->streams, n->index);
-        }
-        else if (n->type == OV_NODE_FPS && n->index >= 0
-                 && n->index < m->nb_fps)
-        {
-            int fi = n->index;
-            bset(out->fps, fi);
-
-            /* Find all stream params of this FPS that match sel_node.
-             * Only meaningful when the selection is a stream.
-             * OR all matching indices into the bitmask. */
-            if (focus == OV_FOCUS_STREAMS
-                && sel_stream_idx >= 0
-                && sel_stream_idx < m->nb_streams)
-            {
-                const char *sname =
-                    m->streams[sel_stream_idx].name;
-                const OV_FPS *f = &m->fps[fi];
-                for (int sp = 0; sp < f->nb_stream_params; sp++)
-                {
-                    if (strcmp(f->stream_param_value[sp],
-                               sname) == 0)
-                    {
-                        out->fps_param_mask[fi] |=
-                            (UINT32_C(1) << sp);
-                    }
-                } /* for sp */
-            } /* if FOCUS_STREAMS */
-
-            (void) fps_is_src; /* suppress unused-var warning */
-        }
-        else if (n->type == OV_NODE_PROC && n->index >= 0
-                 && n->index < m->nb_procs)
-        {
-            bset(out->procs, n->index);
-            if (is_write)
-            {
-                bset(out->proc_writes, n->index);
-            }
-        }
-    } /* for ei */
-}
-
-const char *render_dtype(uint8_t dt)
-{
-    switch (dt)
-    {
-    case _DATATYPE_UINT8:   return "UI8";
-    case _DATATYPE_INT8:    return "SI8";
-    case _DATATYPE_UINT16:  return "U16";
-    case _DATATYPE_INT16:   return "S16";
-    case _DATATYPE_UINT32:  return "U32";
-    case _DATATYPE_INT32:   return "S32";
-    case _DATATYPE_UINT64:  return "U64";
-    case _DATATYPE_INT64:   return "S64";
-    case _DATATYPE_FLOAT:   return "F32";
-    case _DATATYPE_DOUBLE:  return "F64";
-    default:                return "???";
-    }
-}
-
-/**
- * dtype_bytesize - bytes per element for a datatype.
- */
-int dtype_bytesize(uint8_t dt)
-{
-    switch (dt)
-    {
-    case _DATATYPE_UINT8:
-    case _DATATYPE_INT8:    return 1;
-    case _DATATYPE_UINT16:
-    case _DATATYPE_INT16:   return 2;
-    case _DATATYPE_UINT32:
-    case _DATATYPE_INT32:
-    case _DATATYPE_FLOAT:   return 4;
-    case _DATATYPE_UINT64:
-    case _DATATYPE_INT64:
-    case _DATATYPE_DOUBLE:  return 8;
-    default:                return 1;
-    }
-}
 
 static const char *view_label(ov_view_t v)
 {
@@ -427,97 +126,6 @@ static const char *view_label(ov_view_t v)
     case OV_VIEW_PROCS:     return "PROC";
     case OV_VIEW_FPS:       return "FPS";
     default:                return "";
-    }
-}
-
-void clear_row(
-    int row,
-    int col,
-    int width,
-    ov_rgb_t bg)
-{
-    ov_buf_pos(row, col);
-    ov_theme_bg(bg);
-    ov_buf_hline(' ', width);
-}
-
-/* Pad the remainder of a panel's interior row */
-void render_pad_spaces(int chars_written, int panel_width)
-{
-    int remain = (panel_width - 2) - chars_written;
-    if (remain > 0)
-    {
-        ov_buf_hline(' ', remain);
-    }
-}
-
-/**
- * render_scroll_indicators - draw ▲N / ▼N on panel borders.
- * @r:         panel rect
- * @scroll:    current scroll offset (first visible index)
- * @max_rows:  visible rows in the panel body
- * @total:     total item count
- * @accent:    panel accent color for the arrow
- *
- * Draws indicators on the top and bottom border lines of
- * the panel to show how many items are hidden above / below.
- */
-void render_scroll_indicators(
-    OV_RECT  r,
-    int      scroll,
-    int      max_rows,
-    int      total,
-    ov_rgb_t accent)
-{
-    int above = scroll;
-    int below = total - scroll - max_rows;
-    if (below < 0)
-    {
-        below = 0;
-    }
-
-    /* Top border: "▲ N more" right-aligned inside border */
-    if (above > 0)
-    {
-        char buf[32];
-        int n = snprintf(buf, sizeof(buf), " ▲%d ", above);
-        /* display width: space + ▲(1col) + digits + space */
-        int dw = 3;
-        {
-            int tmp = above;
-            while (tmp > 0) { dw++; tmp /= 10; }
-        }
-        int col = r.col + r.width - dw - 2;
-        if (col > r.col + 2)
-        {
-            ov_buf_pos(r.row, col);
-            ov_theme_fg(accent);
-            ov_theme_bg(OV_BG_PANEL);
-            ov_buf_printf("%s", buf);
-            (void) n;
-        }
-    }
-
-    /* Bottom border: "▼ N more" right-aligned inside border */
-    if (below > 0)
-    {
-        char buf[32];
-        int n = snprintf(buf, sizeof(buf), " ▼%d ", below);
-        int dw = 3;
-        {
-            int tmp = below;
-            while (tmp > 0) { dw++; tmp /= 10; }
-        }
-        int brow = r.row + r.height - 1;
-        int col  = r.col + r.width - dw - 2;
-        if (col > r.col + 2)
-        {
-            ov_buf_pos(brow, col);
-            ov_theme_fg(accent);
-            ov_theme_bg(OV_BG_PANEL);
-            ov_buf_printf("%s", buf);
-            (void) n;
-        }
     }
 }
